Accepted all versions from master branch, except for 10.6 SDK setting

Conflicts:
	.gitignore
	SUHost.m
	Sparkle.xcodeproj/project.pbxproj
	is.lproj/Sparkle.strings
	pt_BR.lproj/Sparkle.strings

+++ SUHost.m
@@ -256,7 +256,7 @@
 	OSErr err3 = Gestalt(gestaltSystemVersionBugFix, &bugfix);
 	if (!err1 && !err2 && !err3)
 	{
-		verStr = [NSString stringWithFormat:@"%d.%d.%d", (int)major, (int)minor, (int)bugfix];
+		verStr = [NSString stringWithFormat:@"%ld.%ld.%ld", (long)major, (long)minor, (long)bugfix];
 	}
 	else
 #endif
